% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{Find_max_below_threshold}
\alias{Find_max_below_threshold}
\title{maximum value not exceeded firstMax}
\usage{
Find_max_below_threshold(mat, threshold)
}
\arguments{
\item{mat}{a matrix}

\item{threshold}{threshold}
}
\value{
information of max_below_threshold
}
\description{
maximum value not exceeded firstMax
}
