% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{getInitJaccardMatrix}
\alias{getInitJaccardMatrix}
\title{Initialize matrix}
\usage{
getInitJaccardMatrix(clusterNum, resMarker)
}
\arguments{
\item{clusterNum}{the number of clusters}

\item{resMarker}{Marker gene list with score}
}
\value{
a matrix of jaccard index
}
\description{
Initialize matrix
}
