% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{getInitState}
\alias{getInitState}
\title{Title}
\usage{
getInitState(resMarker, clusterNum, tmp, min.score = 0)
}
\arguments{
\item{resMarker}{Marker gene list with score}

\item{clusterNum}{the number of clusters}

\item{tmp}{the status of each cluster}

\item{min.score}{the minimum of score}
}
\value{
initial state of each cluster
}
\description{
Title
}
