% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{getInitWeightedJaccardMatrix}
\alias{getInitWeightedJaccardMatrix}
\title{Initialize matrix}
\usage{
getInitWeightedJaccardMatrix(clusterNum, resMarker)
}
\arguments{
\item{clusterNum}{the number of clusters}

\item{resMarker}{Marker gene list with score}
}
\value{
a
}
\description{
Initialize matrix
}
