% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{getJaccard}
\alias{getJaccard}
\title{function of Jaccard index}
\usage{
getJaccard(cluster_top, a, b)
}
\arguments{
\item{cluster_top}{Marker gene list}

\item{a}{a cluster to calculate}

\item{b}{a cluster to calculate}
}
\value{
Jaccard index
}
\description{
function of Jaccard index
}
