% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{getSpecificity_score}
\alias{getSpecificity_score}
\title{Calculate marker specificity scores}
\usage{
getSpecificity_score(
  markers_top,
  min.pct.1 = 0.1,
  min.diff = 0,
  min.avg_log2FC = 0.5
)
}
\arguments{
\item{markers_top}{marker genes per cluster.}

\item{min.pct.1}{Minimum percent of cells expressing the gene in the cluster.}

\item{min.diff}{Minimum difference in expression percent between clusters.}

\item{min.avg_log2FC}{Minimum average log2 fold change for marker selection.}
}
\value{
return a specificity score
}
\description{
Calculate marker specificity scores.
}
