% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{getWeightedJaccard}
\alias{getWeightedJaccard}
\title{scSubAnn's weighted Jaccard}
\usage{
getWeightedJaccard(cluster_top, a, b)
}
\arguments{
\item{b}{}
}
\description{
scSubAnn's weighted Jaccard
}
