% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{mergeSteps}
\alias{mergeSteps}
\title{Title}
\usage{
mergeSteps(mergedNodes, combined.min, combined.max)
}
\arguments{
\item{mergedNodes}{recording merged clustering result}

\item{combined.min}{Index of the node to merge(min)}

\item{combined.max}{Index of the node to merge(max)}
}
\description{
Title
}
