% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{setClulterInf}
\alias{setClulterInf}
\title{Title}
\usage{
setClulterInf(resMarker, mergedNodes, resolution)
}
\arguments{
\item{resMarker}{marker gene list with score}

\item{mergedNodes}{recording merged clustering result}

\item{resolution}{resolution}
}
\value{
a tibble of results
}
\description{
Title
}
