% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{Subtypist_Dimplot}
\alias{Subtypist_Dimplot}
\title{Dimensional plot for Subtypist results}
\usage{
Subtypist_Dimplot(
  object = NULL,
  result.table = NULL,
  results.list = NULL,
  reduction = NULL,
  resolution = NULL,
  show = NULL,
  label = FALSE,
  label.size = 4,
  label.color = "black",
  label.box = FALSE,
  repel = FALSE,
  cells.highlight = NULL,
  cols.highlight = "#DE2D26",
  sizes.highlight = 1,
  na.value = "grey50",
  ncol = NULL,
  combine = TRUE,
  raster = NULL,
  raster.dpi = c(512, 512),
  prefix = "Subtypist",
  suffix = NULL,
  select_index = NULL,
  meta.prefix = "phenotypic molecules_"
)
}
\arguments{
\item{object}{A Seurat object.}

\item{result.table}{A data frame containing annotation results.}

\item{results.list}{Subtypist result list}

\item{reduction}{Name of the dimensionality reduction to use.}

\item{resolution}{A vector of resolution values for which to retrieve annotation columns.}

\item{show}{Type of annotation to visualize. One of: 'Result_clusterIndex', 'initial.clusterIndex', 'molecular_phenotype'.}

\item{label}{Whether to label clusters on the plot.}

\item{label.size}{Size of the cluster labels.}

\item{label.color}{Color of the cluster labels.}

\item{label.box}{Whether to draw a box around cluster labels.}

\item{repel}{Whether to repel text labels.}

\item{cells.highlight}{Cells to highlight on the plot.}

\item{cols.highlight}{Color(s) for highlighted cells.}

\item{sizes.highlight}{Size(s) for highlighted cells.}

\item{na.value}{Color for NA values.}

\item{ncol}{Number of columns if combining multiple plots.}

\item{combine}{Whether to combine plots.}

\item{raster}{Whether to use rasterized plotting.}

\item{raster.dpi}{DPI for rasterized output.}

\item{prefix}{Prefix for new metadata columns. Default is \code{"Subtypist"}.}

\item{suffix}{A character suffix to append to the new annotation column names}

\item{select_index}{A named integer vector specifying which phenotype to select for each class (merge_cluster). Names should match cluster IDs in the \code{merge_cluster} column of the result table.}
}
\value{
A ggplot2 object.
}
\description{
Dimensional plot for Subtypist results.
}
\examples{
Subtypist_Dimplot(object = res[[1]], result.table = res[[2]], resolution = c(0.1), show = "molecular_phenotype")
}
