% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{updateDistanceMatrix}
\alias{updateDistanceMatrix}
\title{Update Distance matrix}
\usage{
updateDistanceMatrix(M, Index.min, Index.max, resMarker, clusterNum, operation)
}
\arguments{
\item{M}{matrix needs to be updated}

\item{Index.min}{index to be updated}

\item{Index.max}{index to be updated}

\item{resMarker}{Marker gene list with score}

\item{clusterNum}{the number of clusters}

\item{operation}{the required function}
}
\value{
distance matrix after updating
}
\description{
Update Distance matrix
}
