% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{updateMarkerlist}
\alias{updateMarkerlist}
\title{Title}
\usage{
updateMarkerlist(
  cluster_top,
  combined.min,
  combined.max,
  newMarkers_top,
  clusterNum
)
}
\arguments{
\item{cluster_top}{marker gene list with score}

\item{combined.min}{index to be updated}

\item{combined.max}{index to be updated}

\item{newMarkers_top}{marker gene list with score of merged cluster}

\item{clusterNum}{the number of clusters}
}
\value{
a new marker list
}
\description{
Title
}
