% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{updateSeuratObj}
\alias{updateSeuratObj}
\title{Title}
\usage{
updateSeuratObj(obj, column, combined.min, combined.max, clusterNum)
}
\arguments{
\item{obj}{a seurat object thats needs to be updated}

\item{column}{the column needs to be updated in meta.data}

\item{combined.min}{index to be updated}

\item{combined.max}{index to be updated}

\item{clusterNum}{the number of clusters}
}
\value{
a new Seurat object after updating
}
\description{
Title
}
