% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_utils.R
\name{updateState}
\alias{updateState}
\title{Title}
\usage{
updateState(tmp, combined.min, combined.max, resMarker)
}
\arguments{
\item{tmp}{state of each cluster}

\item{combined.min}{index merged in the previous step}

\item{combined.max}{index merged in the previous step}

\item{resMarker}{Marker gene list}
}
\value{
state of cluster after updating
}
\description{
Title
}
