% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtypist.R
\name{AddSubtypist}
\alias{AddSubtypist}
\title{Title Add Subtypist Annotations to a Seurat Object}
\usage{
AddSubtypist(
  object = NULL,
  resolution = NULL,
  result.table = NULL,
  result.list = NULL,
  prefix = "Subtypist",
  suffix = NULL,
  select_index = NULL,
  meta.prefix = "molecular_phenotype_",
  value.suffix = NULL
)
}
\arguments{
\item{object}{A Seurat object.}

\item{resolution}{A vector of resolution values at which to assign annotations.}

\item{result.table}{A data frame containing Subtypist results with \code{resolution}, \code{merge_cluster}, and \code{molecular_phenotype}.}

\item{result.list}{result list from Subtypist_merge}

\item{prefix}{Prefix for new metadata columns. Default is \code{"Subtypist"}.}

\item{suffix}{A character suffix to append to the new annotation column names}

\item{select_index}{A named integer vector specifying which phenotype to select for each class (merge_cluster). Names should match cluster IDs in the \code{merge_cluster} column of the result table.}

\item{meta.prefix}{The prefix used for new column names added to the metadata.}

\item{value.suffix=}{The suffix used for added to the new columns}
}
\value{
A Seurat object with new metadata columns for Subtypist annotations.
}
\description{
Title Add Subtypist Annotations to a Seurat Object
}
\examples{
library(Seurat)
res.table = tibble::tibble(resolution =c(0.1,0.1,0.1),merge_cluster = c(0,1,2),initial_cluster = c(0,1,2),molecular_phenotype = c('G1','G2','G3'),Score = c(1,2,4))
object <- AddSubtypist(object =res[[1]],result.table=res.table,resolution = c(0.1))
}
