% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtypist.R
\name{saveResults}
\alias{saveResults}
\title{Title}
\usage{
saveResults(result.table, path, name)
}
\arguments{
\item{result.table}{results A data.frame or tibble containing scSubAnn results.}

\item{path}{Directory path to save the file.}

\item{name}{File name, should end with ".csv", ".tsv", or ".xlsx".}
}
\value{
The processed results (invisibly).
}
\description{
Title
}
\examples{
saveResults(res[[2]], path = "output", name = "subtype.csv")
saveResults(res[[2]], path = "output", name = "subtype.xlsx")
}
