% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtypist.R
\name{sortScore}
\alias{sortScore}
\title{Title Sort scoring results by resolution}
\usage{
sortScore(result.table, .f = mean)
}
\arguments{
\item{result.table}{results  data frame.}

\item{.f}{A function to summarize \code{Score} by \code{resolution} (default: \code{mean}).}
}
\value{
a sorted tibble with \code{resolution} and summarized \code{value}.
}
\description{
Title Sort scoring results by resolution
}
\examples{
sortScore(res[[2]],mean)
}
