% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtypist.R
\name{Subtypist_merge}
\alias{Subtypist_merge}
\title{Title}
\usage{
Subtypist_merge(
  object,
  min.resolution = 0.3,
  max.resolution = 1.5,
  by = 0.1,
  max.steps = 100,
  use.assay = "RNA",
  cluster_assay = "integrated",
  n.top = 500,
  min.pct.1 = 0.1,
  min.diff = 0.1,
  min.avg_log2FC = 0.5,
  logfc.threshold = 0.1,
  prefix = "Subtypist"
)
}
\arguments{
\item{object}{a Seurat object}

\item{min.resolution}{the minimum value of resolution}

\item{max.resolution}{the maximum value of resolution}

\item{by}{increment of the sequence.}

\item{use.assay}{Name of assay to use}

\item{n.top}{the number of top genes}

\item{min.pct.1}{the minimum value of pct.1}

\item{min.diff}{the minimum difference between pct.1 and pct.2}

\item{logfc.threshold}{the threshold of the log2 Flod Change}

\item{prefix}{String prefix for naming metadata columns. Each column name will be constructed as "\if{html}{\out{<prefix>}}.snn_res.\if{html}{\out{<resolution>}}"}

\item{cluster.assay}{Name of the assay in the Seurat object to use for clustering}
}
\value{
list of objects and merged results
}
\description{
Title
}
\examples{
Subtypist_merge(object=Seu,min.resolution=0.1,max.resolution=0.2,by=0.1,use.assay="RNA",cluster_assay = "RNA")
}
